requirements.txt
# requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------

matplotlib>=3.2.2
numpy>=1.18.5
numpy-stl>=2.11.2
vtk>=8.2.0
vtkplotlib>=0.2.0
opencv-python>=4.1.1
albumentations>=0.4.6

tensorflow>=2.11.0
flask>=1.1.2

pandas>=1.1.4
seaborn>=0.11.0
